Vanw

+++ banking/src/main/java/com/project88/banking/repository/IUserRepository.java
@@ -30,7 +30,7 @@ public interface IUserRepository extends JpaRepository<User, Long> {
     User findByUsername(@Param("username") String username);
 
     @Query("SELECT u FROM User u WHERE u.userID = :userId")
-    User getUserById(@Param("userId") short userId);
+    User getUserById(@Param("userId") Long userId);
 
     @Query("SELECT new com.project88.banking.dto.GetAllUserDTO(u.userID, CONCAT(u.firstName, ' ', u.lastName), u.email, u.phone, c.cardNumber) FROM User u JOIN u.cardNumber c ")
     Page<GetAllUserDTO> findAllUsers(Pageable pageable);
